Test & refactor the newBundle logic
Refactor to pass the target directory as a configuration option.

The configuration is typed as an object that could later on have
other fields, to eg write things to an s3 instance.

Pass the out directory from the main app in `app.ts`.

Write unit tests exercising the `newBundle` logic.